requirements.txt
azure-ai-ml
numpy
pandas
scikit-learn
matplotlib
mlflow
azureml-mlflow